Adding a downnloaded .gitignore
Source of downloaded .gitignore file: https://github.com/github/gitignore/blob/main/Unity.gitignore

This .gitignore file is downloaded because it is recommended by most to have a .gitignore file to be set up like this for Unity projects. Also, I had to add an additional file that is not hidden or else GitHub won't commit this change/addition of .gitignore file.