90% of Task #284 completed.
Just completed the Fast Travel functionality. For the remainder of this task, I want to provide a brief but thorough documentation of the Fast Travel function so the whole team can fully understand how it works and I will provide that documentation in PDF file in my next commit tomorrow.